Adjusted data types in models to be strictly correct per spec, including treating "long" in the spec as 32-bit integers Implemented IPlugin.ValidateDataOnCard, including validation on each model to test for required fields, length, etc. Added backing integer variables for each enumeration in model objects to handle/validate invalid values

+++ ISOv4Plugin/Mappers/AllocationStampMapper.cs
@@ -61,7 +61,7 @@ namespace AgGateway.ADAPT.ISOv4Plugin.Mappers
             ISOPosition position = new ISOPosition();
             position.PositionEast = (decimal)location.Position.X;
             position.PositionNorth = (decimal)location.Position.Y;
-            position.PositionUp = (long)location.Position.Z;
+            position.PositionUp = (int)location.Position.Z;
             if (location.GpsSource != null)
             {
                 position.NumberOfSatellites = (byte?)location.GpsSource.NumberOfSatellites;
